Move commands/edit utils package up to commands
