+++ pkg/commands/edit/add/addmetadata_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package add
 
 import (
-	"reflect"
 	"testing"
 
 	"sigs.k8s.io/kustomize/v3/pkg/commands/kustfile"
@@ -360,38 +359,5 @@ func TestAddLabelForce(t *testing.T) {
 	v.VerifyCall()
 	if err != nil {
 		t.Errorf("unexpected error: %v", err.Error())
-	}
-}
-
-func TestConvertToMap(t *testing.T) {
-	var o addMetadataOptions
-	args := "a:b,c:\"d\",e:\"f:g\",g:h:k"
-	expected := make(map[string]string)
-	expected["a"] = "b"
-	expected["c"] = "d"
-	expected["e"] = "f:g"
-	expected["g"] = "h:k"
-
-	result, err := o.convertToMap(args)
-	if err != nil {
-		t.Errorf("unexpected error: %v", err.Error())
-	}
-
-	eq := reflect.DeepEqual(expected, result)
-	if !eq {
-		t.Errorf("Converted map does not match expected, expected: %v, result: %v\n", expected, result)
-	}
-}
-
-func TestConvertToMapError(t *testing.T) {
-	var o addMetadataOptions
-	args := "a:b,c:\"d\",:f:g"
-
-	_, err := o.convertToMap(args)
-	if err == nil {
-		t.Errorf("expected an error")
-	}
-	if err.Error() != "invalid annotation: ':f:g' (need k:v pair where v may be quoted)" {
-		t.Errorf("incorrect error: %v", err.Error())
 	}
 }
